Do not require an IPv4 address to use ethernet devices
`pcap_lookupnet()` succeeds only if we have an IPv4 address on the
interface. This call was added to get a netmask that would be in turn
used by `pcap_compile()`. However, this netmask is only needed to test
for IPv4 broadcast addresses. This is a niche use. Just do not provide a
netmask. Being able to "bridge" an interface without an IP address is
more important. `tcpdump` doesn't provide a netmask to `pcap_compile()`
either.